requirements.txt
zeroconf
aiohttp
cryptography
pytest
pytest-asyncio
pytest-homeassistant-custom-component
